Transfers: handle multi-source jobs in receiver. #704
The receiver registers to the "Completion" message sent by fts to the
activemq. This message is generated at the end of a file transfer
(both for success and fail). However, the job can be incomplete yet
(multi-hop transfers; multi-source transfer in which one source failed;
a bulk transfer which has other independent transfers to be done).

Ignore Active jobs unless they are bulk transfers (job_m_replica==True).
In such a way, source failures in multi-source transfers are ignored.
For bulk transfers, each completion is analyzed separately.
Multi-hop is mostly handled because the job_status for multihop is
always set to "UNKNOWN" by fts, but cases of partial failures still
have to be left to be handled by the poller.

Add a couple of receiver tests. Had to add an additional argument to
the receiver because of "vo" mismatch in our test environment: the FTS
vo is set from the CN of the client certificate for API calls to FTS,
which is `RucioClient` and is different from the one on the rucio side
(atlas).