Transfers: rework Topology
Topology is now a subclass of RseCollection and implements a graph
data structure. A sub-class of RseData (Node) and an "Edge" class
where introduced to represent the connection between RSEs.

Because of a strong link between multiple Nodes within the same topology
(via edges), nodes stored in the rse_id_to_data_map member now cannot be
created outside the topology class anymore. All the conversion between
an RSE_ID towards a Node now must be done by the topology class itself
to preserve the consistency, hence the light rework of the
`list_transfer_requests_and_source_replicas` function. This allowed
to get rid of the cryptic, difficult to understand "setdefault"
operation when constructing shortest paths.

Shortest path computations now directly traverse the graph of Node-Edges
without having to convert back-and-forth between rse_id and RseData
objects.

Because the graph is loaded lazily, i.e., not all nodes and edges are
always loaded, we have to maintain a node attribute which allows to know
the status of the edge load.

Under normal operation, nothing changes for now, but this will
potentially allow to pre-load the entire graph and perform many path
computations on that graph without requiring database access.

While performing the change it became clear that fetching unavailable
and restricted RSEs upfront and storing them inside the topology class
only complicates the code. Directly use the columns of the RSE table
and the attributes (which are already fetched anyway) whenever needed.