Core & Internals: annotate and rework config retrieval
The change has a functional impact. Before this change, core/config
was automatically translating strings to bool, int or float; but
common/config was not performing the same translation for data
originating from the config file.
As a consequence, a value 'no' would be returned as False if retrieved
from the database, but was returned as 'no' if retrieved from the file
using config_get.
At the same time, a value '0' would be returned as True if retrieved
from the database, while returned as False if retrieved from the file
by config_get_bool.

This change ensures the same behavior is respected for both sources
of configuration. config_get now always return a string. If a different
type is needed, the corresponding config_get_[int,float,bool,list]
function must be called. Bool translation now always respects the same
rules.

A follow-up commit will substitute config_get all over the code with
the correct config_get_[int,float,bool,list] function (where needed).