Transfers: only use one topology per submitter and preparer
As the topology object represents a structured graph, and maintaining
that structure is more costly, share the topology object across
threads.

Protect the modification of the internal structure using a mutex.

Use a wrapper class to maintain the lifecycle of the topology and
re-create a new object at regular intervals. This is a poor man's
solution for re-fetching fresh data about RSEs and distances from
the database.