WIP: Transfers: prototype bittorrent transfertool using deluge
Additional metadata is added to each file did containing the
merkle hash tree of the bittorrent v2 format. This way, we can
reconstruct the .torrent files from this data, allowing us to
transfer files directly between RSEs using the 'deluge' bittorrent
clients running on each of the RSEs.

Deluge is client/server and is written in python. Because of that,
it seemed like a very good idea to rely on this bittorrent client
for integration with rucio. However, it is very heavily async and
relies on twisted (which I don't know at all) for its internals.
Because of that, the current prototype implementation using deluge
is overly complicated and hacky. It spawns a separate process for
each api call to deluge because the twisted async loop is not
restartable. There should definitely be better ways to handle this...