Tests: initial token infrastructure
Install and configure two oidc providers: indigo iam and keycloak.
For this, had to rework the container initialization to wait for the
database to be ready and to initialize it with the test configuration.

Configure xrood in the XRD2 container to allow token authentication:
this required a separate entrypoint and mounting multiple configuration
files into the container.

Add two Webdav Containers. Only WEB0 is configured for token support,
the second one relies on certificates.