Tests: add test transfer with tokens
Add required clients and scopes to indigoiam. I didn't try to make it
work with keycloak yet. Configure xrootd to accept webdav requests,
but I doubt it actually performs any authentication/authorisation
of the tokens which are passed to it.

Add a test which submits a transfer and verifies that the 'oauth2'
authentication was actually used by fts.